+++ CHANGELOG.md
@@ -5,6 +5,10 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Fixed
+
+- remove name field from upcloud_zone data source
+
 ## [5.20.3] - 2025-03-06
 
 ### Fixed